# backend/openapi.yaml
openapi: 3.0.3
info:
  title: GigLink Marketplace API
  version: 1.0.0
  description: >
    A scoped-down Fiverr-like marketplace without payments.
    Users register as buyers or sellers. Sellers create gigs. Buyers place orders/requests.
    Sellers accept/reject/complete orders. Buyers can leave reviews.

servers:
  - url: http://localhost:8000

tags:
  - name: Auth
  - name: Users
  - name: Gigs
  - name: Orders
  - name: Reviews

paths:
  # ---------------------------
  # AUTH
  # ---------------------------
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user (buyer or seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or duplicate email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------
  # USERS
  # ---------------------------
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get a user profile by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Users]
      summary: Update user profile (must be the same user)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessUserResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (updating another user)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Users]
      summary: Delete user account (must be the same user)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (deleting another user)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------
  # GIGS
  # ---------------------------
  /api/gigs:
    get:
      tags: [Gigs]
      summary: Browse/search gigs
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: category
          in: query
          required: false
          schema: { type: string }
        - name: minPrice
          in: query
          required: false
          schema: { type: number, format: float, minimum: 0 }
        - name: maxPrice
          in: query
          required: false
          schema: { type: number, format: float, minimum: 0 }
      responses:
        "200":
          description: List of gigs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GigPublic"

    post:
      tags: [Gigs]
      summary: Create a gig (seller only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GigCreate"
      responses:
        "201":
          description: Gig created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessGigResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (buyer cannot create gigs)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/gigs/{id}:
    get:
      tags: [Gigs]
      summary: Get gig details by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Gig details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GigPublic"
        "404":
          description: Gig not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Gigs]
      summary: Update a gig (seller only, must own gig)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GigUpdate"
      responses:
        "200":
          description: Gig updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessGigResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (not owner or wrong role)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Gig not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Gigs]
      summary: Delete a gig (seller only, must own gig)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Gig deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (not owner or wrong role)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------
  # ORDERS
  # ---------------------------
  /api/orders:
    post:
      tags: [Orders]
      summary: Create an order/request for a gig (buyer only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessOrderResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (seller cannot create orders)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Gig not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Orders]
      summary: List orders for logged-in user (buyers placed, sellers received)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderPublic"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get a single order by id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPublic"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Orders]
      summary: Update order status (seller only; accept/reject/complete)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessOrderResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (buyer cannot change status / not owner)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Orders]
      summary: Cancel an order (buyer only; typically only if pending)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (seller cannot cancel / not allowed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------
  # REVIEWS
  # ---------------------------
  /api/reviews:
    post:
      tags: [Reviews]
      summary: Create a review (buyer only; typically after a completed order)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreate"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessReviewResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (seller cannot create reviews)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Gig not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/gigs/{id}/reviews:
    get:
      tags: [Reviews]
      summary: Get all reviews for a gig
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of reviews for the gig
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewPublic"

  /api/reviews/{id}:
    delete:
      tags: [Reviews]
      summary: Delete a review (owner/admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Review deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (not review owner)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # ---------------------------
    # ENUMS
    # ---------------------------
    UserRole:
      type: string
      enum: [buyer, seller]

    OrderStatus:
      type: string
      enum: [pending, accepted, rejected, completed]

    # ---------------------------
    # AUTH SCHEMAS
    # ---------------------------
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        role:
          $ref: "#/components/schemas/UserRole"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }

    AuthResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        token: { type: string, example: "jwt.token.here" }
        user:
          $ref: "#/components/schemas/UserPublic"

    # ---------------------------
    # USER SCHEMAS
    # ---------------------------
    UserPublic:
      type: object
      properties:
        id: { type: string, example: "b3d1c3a2-..." }
        name: { type: string }
        email: { type: string, format: email }
        role:
          $ref: "#/components/schemas/UserRole"
        bio: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    UserUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        bio: { type: string }

    SuccessUserResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        user:
          $ref: "#/components/schemas/UserPublic"

    # ---------------------------
    # GIG SCHEMAS
    # ---------------------------
    GigCreate:
      type: object
      required: [title, description, price, category]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        price: { type: number, format: float, minimum: 0 }
        category: { type: string, minLength: 1 }
        tags:
          type: array
          items: { type: string }

    GigUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        price: { type: number, format: float, minimum: 0 }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
        is_active: { type: boolean }

    GigPublic:
      type: object
      properties:
        id: { type: string }
        seller_id: { type: string }
        title: { type: string }
        description: { type: string }
        price: { type: number, format: float }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    SuccessGigResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        gig:
          $ref: "#/components/schemas/GigPublic"

    # ---------------------------
    # ORDER SCHEMAS
    # ---------------------------
    OrderCreate:
      type: object
      required: [gigId, message]
      properties:
        gigId: { type: string }
        message: { type: string, minLength: 1 }

    OrderUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"

    OrderPublic:
      type: object
      properties:
        id: { type: string }
        gig_id: { type: string }
        buyer_id: { type: string }
        message: { type: string }
        status:
          $ref: "#/components/schemas/OrderStatus"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    SuccessOrderResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        order:
          $ref: "#/components/schemas/OrderPublic"

    # ---------------------------
    # REVIEW SCHEMAS
    # ---------------------------
    ReviewCreate:
      type: object
      required: [gigId, rating]
      properties:
        gigId: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }

    ReviewPublic:
      type: object
      properties:
        id: { type: string }
        gig_id: { type: string }
        buyer_id: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }
        created_at: { type: string, format: date-time }

    SuccessReviewResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        review:
          $ref: "#/components/schemas/ReviewPublic"

    # ---------------------------
    # GENERIC RESPONSES
    # ---------------------------
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: "OK" }

    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string, example: "Something went wrong" }
            details:
              type: object
              additionalProperties: true